+++ test/unit/mocks.mjs
@@ -3,6 +3,7 @@
  * @typedef {import('../../src/typedefs.mjs').JQueryAjaxSettings} JQueryAjaxSettings
  * @typedef {import('../../src/typedefs.mjs').Deferred} Deferred
  * @typedef {import('../../src/typedefs.mjs').AsyncComplete} AsyncComplete
+ * @typedef {import('../../src/typedefs.mjs').AnyType} AnyType
  */
 
 /**
@@ -11,8 +12,27 @@
  */
 export function MockDOMParser() {
     return {
-        parseFromString: () => {
-            return { namespaceURI: 'http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul' }
+        /**
+         * Mock out parsing an XML string to an object
+         * @param {string} xmlString A string of XML
+         * @returns {object} The parsed XML document object (or <parsererror>Parse Error</parsererror>)
+         */
+        parseFromString: (xmlString) => {
+            if (/parsererror/.test(xmlString)) {
+                return {
+                    namespaceURI: 'http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul',
+                    parsererror: 'There was a parser error'
+                }
+            } else if (/notxml/.test(xmlString)) {
+                return {}
+            } else {
+                return {
+                    namespaceURI: 'http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul',
+                    root: {
+                        item: 'test'
+                    }
+                }
+            }
         }
     }
 }
@@ -58,32 +78,36 @@ export function createMockXHR(handler={}) {
 export function Deferred(xhr) {
     return {
         isResolved: false,
-        data: null,
-        _complete: null,
+        _context: null,
+        _data: null,
+        _always: null,
         resolve: function() {
-            this.resolveWith(xhr)
+            this.resolveWith({}, [xhr])
         },
         /**
          * Creates a mock Deferred.resolveWith() for use in tests
+         * @param {object} context The context to use for callbacks
+         * @param {AnyType[]} args An array of the arguments to pass into the callbacks
          * @returns {boolean} Always returns true
          */
-        resolveWith: function(data) {
-            this.data = data
+        resolveWith: function(context, args=[]) {
+            this._data = args
+            this._context = context || {}
             this.isResolved = true
-            if (this._complete) {
-                this._complete(this.data)
+            if (this._always) {
+                this._always.call(this._context, ...this._data)
             }
         },
 
         /**
-         * Creates a mock Deferred.complete() for use in tests
-         * @param {AsyncComplete} callback The function to call once the Deferred is complete
+         * Creates a mock Deferred.always() for use in tests
+         * @param {AsyncComplete} callback The function to call once the Deferred is complete (resolved or rejected)
          * @returns {void}
          */
-        complete: function (callback) {
-            this._complete = callback
-            if (this.isResolved && this._complete) {
-                this._complete(this.data)
+        always: function (callback) {
+            this._always = callback
+            if (this.isResolved && this._always) {
+                this._always.call(this._context, ...this._data)
             }
         }
     }
@@ -97,12 +121,15 @@ export function Deferred(xhr) {
  * @returns {void}
  */
 export function getJQueryMock() {
-    const mockJQuery = function jQuery(selector) {
-        if (selector === 'parseerror') {
-            // see xhr.mjs -> parseXML
-            return []
+    const mockJQuery = function jQuery(selector, parent) {
+        if (selector === 'parsererror') {
+            if (/parsererror/.test(JSON.stringify(parent))) {
+                return [JSON.stringify(parent)]
+            } else {
+                return []
+            }
         } else {
-            return makeNodeSelection()
+            return makeNodeSelection(selector)
         }
     }
     mockJQuery.isMock = true
@@ -132,7 +159,12 @@ export function ajax(url, settings={}) {
         settings.url = url
     }
     
-    const xhr = (settings.xhr && settings.xhr()) || createMockXHR({ url: settings.url })
+    const xhr = (settings.xhr && settings.xhr()) || createMockXHR({ url: settings.url || '' })
+    
+    if (typeof settings.complete === 'function') {
+        settings.complete(xhr)
+    }
+
     return new Deferred(xhr)
 }
 
@@ -156,26 +188,30 @@ function ajaxSetup(...settings) {
 
 /**
  * Creates a mock selection object a la $('selector')
+ * @param {object} elem The selected element
  * @returns {boolean} Always returns true
  */
-function makeNodeSelection() {
-    return { trigger, text, length: 1 }
+function makeNodeSelection(elem) {
+    elem = (typeof elem === 'string') ? { type: elem } : elem
+    return { 0: elem, trigger, text, length: 1 }
 }
 
 /**
  * A mock $.globalEval() for use in tests
+ * @param {string} script The script to eval
  * @returns {boolean} Always returns true
  */
-function globalEval() {
-    return true
+function globalEval(script) {
+    global.scriptEval = script
 }
 
 /**
  * A mock $.isXMLDoc() for use in tests
+ * @param {object} object A possible XML document
  * @returns {boolean} Always returns true
  */
-function isXMLDoc() {
-    return true
+function isXMLDoc(object) {
+    return object && typeof object === 'object' && object.namespaceURI
 }
 
 /**
@@ -191,5 +227,5 @@ function trigger() {
  * @returns {boolean} Always returns true
  */
 function text() {
-    return 'text'
+    return (this && this[0]) ? JSON.stringify(this[0]) : '(empty)'
 }